Setup mssql example sql as part of integration-env
